Added most-used image functions to `image` module
Also moved `save_image` from `image` module to `image.utils`
